+++ pyupgw/_helpers.py
@@ -1,6 +1,8 @@
 """Helper functions"""
 
 import asyncio
+import base64
+import pickle
 from collections.abc import Callable
 
 
@@ -20,3 +22,18 @@ async def async_future_helper(
     """
     future = getter(await asyncio.to_thread(func, *args))
     return await asyncio.wrap_future(future)
+
+
+class LazyEncode:
+    """Lazily encode an object for logging"""
+
+    def __init__(self, obj):
+        self._obj = obj
+
+    def __str__(self):
+        """Return ``self.encode()``"""
+        return self.encode()
+
+    def encode(self):
+        """Pickle and base64 encode the wrapped object"""
+        return base64.b64encode(pickle.dumps(self._obj)).decode()